# backendmemoire/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teranga GESCRIM API",
        "description": "API REST pour le système de Gestion de la Criminalité et de la Délinquance — Direction de la Sécurité Publique (DSP) du Sénégal.",
        "contact": {
            "name": "NAMASTECH",
            "email": "admin@gescrim.sn"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://gescrim.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/accidents": {
            "get": {
                "tags": [
                    "Accidents"
                ],
                "summary": "Liste des accidents de la circulation",
                "operationId": "ea799341a618789de80410a54dc1bfb1",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "mat\\u00e9riel",
                                "corporel",
                                "mortel"
                            ]
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commune_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accidents"
                ],
                "summary": "Enregistrer un accident",
                "operationId": "f18dbf00ffd5970d472428a991d61d86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "date",
                                    "lieu",
                                    "commune_id",
                                    "service_id"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "corporel",
                                        "enum": [
                                            "matériel",
                                            "corporel",
                                            "mortel"
                                        ]
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-12"
                                    },
                                    "lieu": {
                                        "type": "string",
                                        "example": "Autoroute à péage, km 15"
                                    },
                                    "commune_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "moyen": {
                                        "type": "string",
                                        "example": "Véhicule"
                                    },
                                    "cause_probable": {
                                        "type": "string",
                                        "example": "Excès de vitesse"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": 14.7228
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": -17.4567
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/infractions": {
            "get": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Liste des infractions",
                "operationId": "6e55860eab55d36fa726adcdd973f5c6",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commune_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "issue",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Constat\\u00e9e",
                                "D\\u00e9f\\u00e9r\\u00e9e"
                            ]
                        }
                    },
                    {
                        "name": "type_infraction_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sync_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "synced"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Infractions"
                ],
                "summary": "Enregistrer une infraction",
                "description": "Crée une nouvelle infraction. Le champ sync_status peut être 'pending' pour les saisies offline.",
                "operationId": "afd0a32797f783bca664f687618d42ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_infraction_id",
                                    "service_id",
                                    "annee",
                                    "date",
                                    "lieu",
                                    "commune_id",
                                    "issue"
                                ],
                                "properties": {
                                    "type_infraction_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-12"
                                    },
                                    "lieu": {
                                        "type": "string",
                                        "example": "Marché Sandaga, Dakar"
                                    },
                                    "commune_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "issue": {
                                        "type": "string",
                                        "example": "Constatée",
                                        "enum": [
                                            "Constatée",
                                            "Déférée"
                                        ]
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": 14.6928
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": -17.4467
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "sync_status": {
                                        "type": "string",
                                        "example": "synced",
                                        "enum": [
                                            "pending",
                                            "synced"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/sync/batch": {
            "post": {
                "tags": [
                    "Synchronisation"
                ],
                "summary": "Synchroniser les données offline par lot",
                "operationId": "17b478c48719a9159824836a063841b9",
                "responses": {
                    "200": {
                        "description": "Synchronisation réussie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sync/status": {
            "get": {
                "tags": [
                    "Synchronisation"
                ],
                "summary": "État de synchronisation",
                "operationId": "7bd5ab03d3512ac159a2922b6e67f5af",
                "responses": {
                    "200": {
                        "description": "Statut de sync"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Erreur rencontrée"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token JWT (sans le mot Bearer)"
            }
        }
    },
...
}